+++ NEWS.md
@@ -25,6 +25,8 @@ Regarding the SWLN template:
 
 New presets/rules:
 
+* 2026/03/24 - `DisableRDCShareLocalDrive`/`EnableRDCShareLocalDrive`/`ViewRDCShareLocalDrive` -> Remote Desktop Client could not share Local Drive
+* 2026/03/24 - `DisableRDCSavePassword`/`EnableRDCSavePassword`/`ViewRDCSavePassword` -> Remote Desktop Client could not save Passwords
 * 2026/03/13 - `ViewFirefoxTLS` -> Firefox TLS Minimum Version
 * 2026/03/10 - `ViewWindowsServicingLevel` -> Windows systems must be maintained
 * 2026/03/10 - `RemoveUserInAdminGroup`/`ViewUserInAdminGroup` -> Limit Users in Admin Group (Use `$Global:SWMB_Custom` key `LocalAdminRegex`)